// generated/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/**
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/
export const UserRole = {
  ADMIN: 'ADMIN',
  CASHIER: 'CASHIER',
  AGENT: 'AGENT'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ClientStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
  BANNED: 'BANNED'
} as const

export type ClientStatus = (typeof ClientStatus)[keyof typeof ClientStatus]


export const TransactionType = {
  CHARGE: 'CHARGE',
  REDEEM: 'REDEEM',
  ADJUSTMENT: 'ADJUSTMENT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const PaymentMethod = {
  CASH: 'CASH',
  DEBIT_CARD: 'DEBIT_CARD',
  CREDIT_CARD: 'CREDIT_CARD',
  BANK_TRANSFER: 'BANK_TRANSFER',
  OTHER: 'OTHER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const ContactChannel = {
  WHATSAPP: 'WHATSAPP',
  INSTAGRAM: 'INSTAGRAM',
  FACEBOOK: 'FACEBOOK',
  TIKTOK: 'TIKTOK',
  PHONE_CALL: 'PHONE_CALL',
  SMS: 'SMS',
  IN_PERSON: 'IN_PERSON',
  OTHER: 'OTHER'
} as const

export type ContactChannel = (typeof ContactChannel)[keyof typeof ContactChannel]


export const ContactDirection = {
  INBOUND: 'INBOUND',
  OUTBOUND: 'OUTBOUND'
} as const

export type ContactDirection = (typeof ContactDirection)[keyof typeof ContactDirection]